requirements.txt
flask==2.0.1
flask-cors==3.0.10
faiss-cpu==1.7.2
numpy==1.21.2
openai==0.27.0
werkzeug==2.0.1
python-dotenv==0.19.0